Clean up SessionArtifact implementation
